+++ h/accounts/events.py
@@ -1,0 +1,4 @@
+class LoginEvent(object):
+    def __init__(self, request, user):
+        self.request = request
+        self.user = user
